軽微な修正

+++ lib/car_control/control/instruction/implementation/stop_instruction.cpp
@@ -8,8 +8,8 @@ int StopInstruction::runCoroutine() {
     COROUTINE_BEGIN();
     start_left_speed_ = left_wheel_->Speed();
     start_right_speed_ = right_wheel_->Speed();
-    left_unit_decelerate_amount_ = start_left_speed_ / duration_millis_;  /// 左車輪の減速量
-    right_unit_decelerate_amount_ = start_right_speed_ / duration_millis_;  /// 右車輪の減速量
+    left_unit_decelerate_amount_ = (double)start_left_speed_ / duration_millis_;  /// 左車輪の減速量
+    right_unit_decelerate_amount_ = (double)start_right_speed_ / duration_millis_;  /// 右車輪の減速量
 
     while (left_wheel_->Speed() > 0 || right_wheel_->Speed() > 0) {
         left_wheel_->Decelerate(left_unit_decelerate_amount_);
